refactor: split org report writers by format
